requirements.txt
fastapi
uvicorn
anthropic
groq
python-dotenv
pymongo
python-multipart
